Changed return type to size_t

+++ 0x16-doubly_linked_lists/1-dlistint_len.c
@@ -8,7 +8,7 @@
  */
 size_t print_dlistint(const dlistint_t *h)
 {
-	int elements = 0;
+	size_t elements = 0;
 
 	while (h != NULL)
 	{
